+++ src/Tests/AllTests.fs
@@ -33,6 +33,7 @@ let allTests =
                 IotHub.tests
                 JsonRegression.tests
                 KeyVault.tests
+                LoadBalancer.tests
                 LogAnalytics.tests
                 Maps.tests
                 NetworkSecurityGroup.tests
@@ -42,6 +43,7 @@ let allTests =
                 Sql.tests
                 StaticWebApp.tests
                 Storage.tests
+                TrafficManager.tests
                 Types.tests
                 VirtualMachine.tests
                 VirtualNetworkGateway.tests